:bricks: ci: Generate Dockerfile and first continuos integration
